/time is no longer a debug command, deserves its own file
